lz?? based compression logic initial skeleton starting
Go thro a string given to the logic and build a dictionary of
uniq char patterns (sub strings) as they appear the 1st time.

Any new uniq char pattern can be considered to consist of
a existing uniq char pattern + the last char in the new uniq char
pattern.

Thus In turn print the new uniq char pattern as a pair consisting
of the index corresponding to the existing uniq char pattern and
the last char in the new uniq char pattern