Added compress_str which calls the bytes based one
the bytes based compress sets its diStart by default to 256, so
that the 1st 256 implicit entries in the dictionary correspond
to binary values from 0 to 255.

the new string based compress_str sets the diStart by default to
128, so that the 1st 128 implicit entries in the dictionary
correspond to binary values from 0 to 127 i.e the base ascii
char set.

Even thou compress_str calles bytes with 'utf-8' as the encoding
of the string, as the diStart by default is set to 128, strictly
speaking it works for ascii strings and not pure utf-8 strings.