Cleanup rename functions to mirror logic properly
As currently compress and decompress functions work on bytearray
and bytes, it can work only with buffer of bytes were the actual
data is less than 255. Rather the maximum value in the byte decides
how much space is available for sequerce of values to be stored
as dictionary elements

So if the buffer contains traditional ascii chars, then the maximum
value that will be there in the buffer will be less than 128. So
0-127 will act as the implicit single value dictionary elements and
128-255 will act as the explicit sequence of single values related
dictionary elements.

Similarly if a buffer contains values always less than say 200,
then 0-199 will correspond to the implicit single value dictionary
elements and 200-255 will act as the explicit sequence of single
values related dictionary elements.

diStart should be 128 for the 1st example and 200 for the 2nd ex

NOTE: If we want to compress/decompress a byte array/buffer which
can contain values upto 255 in each byte of the buffer, then we
will have to Either

a) use a uint9_t or .. or .. or uint16 array to store the
compressed buffer

OR

b) convert the input byte array into a base64 or base128 data
and then use the existing compres_Xin8 or decompress_Xin8 funcs