Handle the bytes at end of the input buffer
As we reach end of the input buffer, we may have data which matches
a existing implict or explicit dictionary element, so which has not
yet been written to the compressed buffer. One is required to handle
this appropriately, so that we dont lose them.

This modification to the code achieves it. However this is bit
inefficient way of doing it. Next commit will give a more simple
and straight forward way of achieving the same.