DictOverFlowHandling P2 - special marker and reset
Made diStart the dictionary index as well as the marker which
indicates that the dictionary index requires to be reset now.

Also the dictionary is reset in a lazy and late manner. i.e Even
thou the logic starts adding elements to the dictionary past the
255th index i.e from 256 onwards, the logic doesnt bother about
reseting the dictionary to a clean state. Because there could be
lot more text in the input data which are not going to be depending
on these beyond the valid dictionary index entries. And thus the
compression can go ahead without problem for compress or decompress
till a point is reached where the input data for compression
depends on a entry beyond 255th dictionary entry. At that moment
the reset marker is written into the compressed byte stream as well
as the dictionary is reset.

The reset marker is written into two adjacent bytes, because each
element in the compressed stream consists of two bytes, the 1st
byte is the index into dictionary and the next byte is the next
char which makes the expanded 1st byte and this byte together a
new unique sequence in the dictionary.

For now I havent taken care of reseting the compress logic to the
proper position in the input data such that compressed data doesnt
lose few chars (which would have corresponded to the byte sequence
which matched the dictionary entry beyond 255th index).