Preparing for a in time dictionary reset
The existing dictionary reset logic is a lazy one, which is fine
for a machine with enough memory. But on a machine with limited
memory, the additional entries beyond 128 explicit entries i.e
starting with index 256 onwards, which are added into dictionary
but discarded (their only use is that, when a match is found
into these forbidden entrires, it triggers the lazy reset logic)
is not feasible.

So require to add a intime reset logic. This would also require
the continue operator to be used. But the existing loop counter
logic which increments at the end of the loop would have made it
inefficient, so the loop counter logic has been updated to inc
at the beginning of the loop.

Also have put some code in place for a in time trigger of dict
reset (but in a commented state), also the input data parsing
position reset logic is not updated properly in this commented
out code.