compress In time dict reset logic 1st part
Have uncommented the basic in time dict reset logic which I had
added last time.

In turn now that while loop counter logic updated to be condusive
to use of continue loop construct (which was done for this purpose
only) the same used in this in time dict reset logic.

As part of the same noticed that as I was reseting the state to
START at the end of the dict update and compressed stream update
code, the continue was bypassing the same and leading to a stale
dictionary entry being added into compressed stream in the immidiate
next loop. This inturn trips up the decompress logic as well as
will lead to corruption. So now the state reset logic is moved
to the begining of this block of code.

NOTE: This code still doesnt handle reseting of the input data
position as required. So few characters that were being processed
when the in time dict reset logic kicked in will be lost.